+++ concrete/codegen/datamod/utils.go
@@ -1,0 +1,56 @@
+// Copyright 2023 The concrete-geth Authors
+//
+// The concrete-geth library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The concrete library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the concrete library. If not, see <http://www.gnu.org/licenses/>.
+
+package datamod
+
+import (
+	"regexp"
+	"strings"
+	"unicode"
+)
+
+func lowerFirstLetter(str string) string {
+	if len(str) == 0 {
+		return ""
+	}
+	runes := []rune(str)
+	runes[0] = unicode.ToLower(runes[0])
+	return string(runes)
+}
+
+func upperFirstLetter(str string) string {
+	if len(str) == 0 {
+		return ""
+	}
+	runes := []rune(str)
+	runes[0] = unicode.ToUpper(runes[0])
+	return string(runes)
+}
+
+func formatTableName(tableName string) string {
+	return upperFirstLetter(tableName)
+}
+
+func formatRowName(tableName string) string {
+	return upperFirstLetter(tableName) + "Row"
+}
+
+func isValidName(name string) bool {
+	if len(name) == 0 {
+		return false
+	}
+	re := regexp.MustCompile(`^[a-zA-Z_][a-zA-Z0-9_]*$`)
+	return re.MatchString(name) && len(strings.TrimSpace(name)) == len(name)
+}
